rewrite recordsources
