+++ db/query.sql
@@ -20,3 +20,6 @@ SELECT * FROM candidate_records WHERE status = 'new' AND id = $1 LIMIT 1;
 
 -- name: SetStatusCandidateRecord :one
 UPDATE candidate_records SET status = $1 WHERE id = $2 RETURNING id;
+
+-- name: GetCandidateRecordBySource :one
+SELECT * FROM candidate_records WHERE source_name = $1 AND source_id = $2 LIMIT 1;
